+++ cyberdrop_dl/crawlers/daftporn.py
@@ -1,0 +1,37 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, ClassVar
+
+from cyberdrop_dl.crawlers.crawler import Crawler, SupportedPaths
+from cyberdrop_dl.url_objects import AbsoluteHttpURL
+from cyberdrop_dl.utils import css
+from cyberdrop_dl.utils.errors import error_handling_wrapper
+
+if TYPE_CHECKING:
+    from cyberdrop_dl.url_objects import ScrapeItem
+
+
+class DaftPornCrawler(Crawler):
+    SUPPORTED_PATHS: ClassVar[SupportedPaths] = {
+        "Video": "/extreme-videos/<slug>",
+    }
+    PRIMARY_URL: ClassVar[AbsoluteHttpURL] = AbsoluteHttpURL("https://www.daftporn.com")
+    DOMAIN: ClassVar[str] = "daftporn"
+
+    async def fetch(self, scrape_item: ScrapeItem) -> None:
+        match scrape_item.url.parts[1:]:
+            case ["extreme-videos", _]:
+                return await self.video(scrape_item)
+            case _:
+                raise ValueError
+
+    @error_handling_wrapper
+    async def video(self, scrape_item: ScrapeItem) -> None:
+        if await self.check_complete_from_referer(scrape_item.url):
+            return
+
+        soup = await self.request_soup(scrape_item.url)
+        name = css.select_text(soup, "title")
+        src = self.parse_url(css.select(soup, "video source", "src"))
+        filename = self.create_custom_filename(name, ext := ".mp4")
+        await self.handle_file(src, scrape_item, name, ext, custom_filename=filename)