+++ tests/crawlers/test_cases/adobe_lightroom.py
@@ -1,0 +1,17 @@
+DOMAIN = "lightroom.adobe"
+TEST_CASES = [
+    {
+        "url": "https://lightroom.adobe.com/shares/2e55bfd877144349b870881fa0f63595",
+        "results": [
+            {
+                "url": "re:https://lightroom.adobe.com/v2c/spaces/2e55bfd877144349b870881fa0f63595/assets/",
+                "debrid_url": None,
+                "referer": "https://lightroom.adobe.com/shares/2e55bfd877144349b870881fa0f63595",
+                "album_id": "338e403f6cdd4971af3bf2e79b676067",
+                "uploaded_at": int,
+                "download_folder": "re:ODFW Commission meeting April 24, 2026 (Adobe Lightroom)",
+            },
+        ],
+        "count": 25,
+    }
+]