+++ tests/crawlers/test_cases/crazyshit.py
@@ -1,0 +1,19 @@
+DOMAIN = "crazyshit"
+TEST_CASES = [
+    {
+        "url": "https://crazyshit.com/cnt/medias/214810-17-people-definitely-having-a-shittier-day-than-you",
+        "results": [
+            {
+                "url": "https://crazyshit.com/cnt/medias/214810-17-people-definitely-having-a-shittier-day-than-you",
+                "filename": "17 PEOPLE DEFINITELY HAVING A SHITTIER DAY THAN YOU [214810].mp4",
+                "debrid_url": "re:https://mediav.crazyshit.com/",
+                "original_filename": "17 PEOPLE DEFINITELY HAVING A SHITTIER DAY THAN YOU",
+                "referer": "https://crazyshit.com/cnt/medias/214810-17-people-definitely-having-a-shittier-day-than-you",
+                "album_id": None,
+                "uploaded_at": None,
+                "download_folder": "re:Loose Files (CrazyShit)",
+            }
+        ],
+        "count": 1,
+    }
+]