+++ tests/crawlers/test_cases/monstercat.py
@@ -1,0 +1,29 @@
+DOMAIN = "monstercat"
+TEST_CASES = [
+    {
+        "url": "https://www.monstercat.com/release/742779563446",
+        "results": [
+            {
+                "url": "https://www.monstercat.com/api/release/af0830b2-0318-4411-b336-e1684774f0f0/track-stream/b381e5aa-3629-497c-b451-388f28a6077b",
+                "filename": "FLUTTER.mp3",
+                "debrid_url": None,
+                "original_filename": "FLUTTER",
+                "referer": "https://www.monstercat.com/release/742779563446",
+                "album_id": "742779563446",
+                "uploaded_at": 1780894800,
+                "download_folder": "re:FLUTTER - STEP OUT (Monstercat)",
+            },
+            {
+                "url": "https://www.monstercat.com/api/release/af0830b2-0318-4411-b336-e1684774f0f0/track-stream/02af32d1-200c-4eab-ae0b-9823bdf4cdfb",
+                "filename": "STEP OUT.mp3",
+                "debrid_url": None,
+                "original_filename": "STEP OUT",
+                "referer": "https://www.monstercat.com/release/742779563446",
+                "album_id": "742779563446",
+                "uploaded_at": 1780894800,
+                "download_folder": "re:FLUTTER - STEP OUT (Monstercat)",
+            },
+        ],
+        "count": 2,
+    }
+]