+++ CHANGELOG.md
@@ -22,7 +22,28 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## UNRELEASED
+## [10.1.0] - 2026-07-10
+
+### Added
+
+- ABStream support
+- Adobe Lightroom support
+- APKMirror support
+- Cara.app support
+- Crazyshit support
+- Daftporn support
+- Dailymotion support
+- GoonBox support
+- Monstercat support
+- Multporn support
+- NaughtyMachinima support
+- Nudedeleted support
+- Pawchive support
+- Pillowcase support
+- Pinterest support
+- Trannygem support
+- Wikifeet support
+- Yourlesbian support
 
 ### Changed
 
@@ -32,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - No videos detected on profiles, playlists or search results (Spankbang)
+- Download URL extraction (Fileditch)
 
 ## [10.0.2] - 2026-07-05